Add envoy gateway streaming support (#377)
* Add reference grant to support httprouting for different namespace

* lint fix

* create reference grant per namespace

* refactor validate routing strategy

* test

* Add streaming support

* nit

* bug fix for streaming

* nit

* nit comment update

* comment model name check from cache till we use static lora

+++ pkg/cache/cache.go
@@ -505,7 +505,7 @@ func parseMetricFromBody(body []byte, metricName string) (float64, error) {
 	return 0, fmt.Errorf("metrics %s not found", metricName)
 }
 
-func (c *Cache) AddRequestTrace(modelName string, inputTokens, outputTokens int) {
+func (c *Cache) AddRequestTrace(modelName string, inputTokens, outputTokens int64) {
 	c.mu.Lock()
 	defer c.mu.Unlock()
